9th commit: Feature- updated crud features of comments

+++ controllers/comment.controller.js
@@ -34,9 +34,55 @@ export const getComments = async (req, res) => {
     }
 };
 
+// EDIT COMMENT
+export const updateComment = async (req, res) => {
+    try {
+        const comment = await CommentModel.findById(req.params.id);
+
+        if (!comment) {
+            return res.status(404).json({
+                message: "Comment not found",
+            });
+        }
+
+        if (comment.user.toString() !== req.user.id) {
+            return res.status(403).json({
+                message: "Unauthorized",
+            })
+
+        }
+
+        comment.text = req.body.text;
+
+        await comment.save();
+
+        res.status(200).json(comment);
+
+    } catch (error) {
+        res.status(500).json({
+            message: error.message,
+        });
+    }
+};
+
 // DELETE COMMENT
 export const deleteComment = async (req, res) => {
     try {
+
+        const comment = await CommentModel.findById(req.params.id);
+
+        if (!comment) {
+            return res.status(404).json({
+                message: "Comment not found",
+            });
+        }
+
+        if (comment.user.toString() !== req.user.id) {
+            return res.status(403).json({
+                message: "Unauthorized",
+            });
+        }
+
         await CommentModel.findByIdAndDelete(
             req.params.commentId
         );
